backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pdfplumber==0.10.3
scikit-learn==1.4.0
numpy==1.26.4

# Optional — better semantic matching (downloads ~80MB model on first use)
# sentence-transformers==2.2.2

# Optional — NLP entity extraction (run: python -m spacy download en_core_web_sm)
# spacy==3.7.2
